requirements.txt
langchain>=0.3.8
langchain-community>=0.3.7
langchain-openai>=0.2.11
streamlit>=1.41.0
pandas>=2.2.3
openpyxl>=3.1.5
python-dotenv>=1.0.1
pytest>=8.3.3
pytest-cov>=5.0.0
ruff>=0.7.4
